article_ist/zenodo_package/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.66.0
bibtexparser>=1.4.0
rispy>=0.7.0
rapidfuzz>=3.6.0
colorama>=0.4.6
anthropic>=0.40.0
